requirements.txt
langchain==0.1.7
langchain-community
openai==1.12.0
python-dotenv==1.0.1 